requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118

numpy==2.1.3
pillow==11.1.0
opencv-python-headless==4.10.0.84
requests==2.32.3
pyautogui==0.9.54
pygetwindow==0.0.9
scikit-learn==1.6.1
timm==1.0.15

torch==2.5.1+cu118
torchvision==0.20.1
torchaudio==2.5.1+cu118

pyinstaller==6.13.0  # 선택: exe 패킹용
